python-mcp-servers/flicker_detection/requirements.txt
# Flicker Detection MCP Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
python-dotenv>=1.0.0
structlog>=24.1.0
# Analysis
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.11.0
# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
